feat: Kline数据计算支持两种模式(fast, accurate)对应分别使用double和Decimal来计算指标数据; 增加BagNum类型适配计算模式; 增加测试用例; 优化exampleUI/Theme

+++ lib/src/indicators/boll.dart
@@ -203,21 +203,21 @@ class BOLLPaintObject<T extends BOLLIndicator> extends SinglePaintObjectBox<T> {
     final children = <TextSpan>[];
 
     final mbTxt = formatNumber(
-      ret.mb,
+      ret.mb.toDecimal(),
       precision: precision,
       cutInvalidZero: true,
       prefix: indicator.mbLabel,
       suffix: ' ',
     );
     final upTxt = formatNumber(
-      ret.up,
+      ret.up.toDecimal(),
       precision: precision,
       cutInvalidZero: true,
       prefix: indicator.upLabel,
       suffix: ' ',
     );
     final dnTxt = formatNumber(
-      ret.dn,
+      ret.dn.toDecimal(),
       precision: precision,
       cutInvalidZero: true,
       prefix: indicator.dnLabel,
